refactored query to use statement
