+++ src/DataMapper/Query/Insert.php
@@ -18,65 +18,41 @@ declare(strict_types=1);
 
 namespace Phalcon\DataMapper\Query;
 
+use PDOStatement;
 use Phalcon\DataMapper\Pdo\Connection;
-
-use function is_int;
+use Phalcon\DataMapper\Pdo\Exception\Exception;
+use Phalcon\DataMapper\Statement\Insert as InsertStatement;
 
 /**
  * Insert Query
  */
-class Insert extends AbstractQuery
+class Insert extends InsertStatement
 {
     /**
-     * Insert constructor.
+     * Create a new instance of this object
      *
+     * @param mixed ...$arguments
+     *
+     * @return static
+     */
+    public static function new(mixed ...$arguments): static
+    {
+        $connection = Connection::new(...$arguments);
+
+        return new static($connection->getDriverName(), $connection);
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param string     $driver
      * @param Connection $connection
-     * @param Bind       $bind
      */
-    public function __construct(Connection $connection, Bind $bind)
-    {
-        parent::__construct($connection, $bind);
-
-        $this->store["FROM"]      = "";
-        $this->store["RETURNING"] = [];
-    }
-
-    /**
-     * Sets a column for the `INSERT` query
-     *
-     * @param string $column
-     *
-     * @return Insert
-     */
-    public function column(string $column, mixed $value = null, int $type = -1): self
-    {
-        $this->store["COLUMNS"][$column] = ":" . $column;
-
-        if (null !== $value) {
-            $this->bind->setValue($column, $value, $type);
-        }
-
-        return $this;
-    }
-
-    /**
-     * Mass sets columns and values for the `INSERT`
-     *
-     * @param array $columns
-     *
-     * @return Insert
-     */
-    public function columns(array $columns): self
-    {
-        foreach ($columns as $column => $value) {
-            if (is_int($column)) {
-                $this->column($value);
-            } else {
-                $this->column($column, $value);
-            }
-        }
-
-        return $this;
+    public function __construct(
+        string $driver,
+        protected Connection $connection
+    ) {
+        parent::__construct($driver);
     }
 
     /**
@@ -92,97 +68,16 @@ class Insert extends AbstractQuery
     }
 
     /**
-     * @return string
-     */
-    public function getStatement(): string
-    {
-        return "INSERT"
-            . $this->buildFlags()
-            . " INTO " . $this->store["FROM"]
-            . $this->buildColumns()
-            . $this->buildReturning();
-    }
-
-    /**
-     * Adds table(s) in the query
+     * Performs a statement in the connection
      *
-     * @param string $table
-     *
-     * @return Insert
+     * @return PDOStatement
+     * @throws Exception
      */
-    public function into(string $table): self
+    public function perform()
     {
-        $this->store["FROM"] = $table;
-
-        return $this;
-    }
-
-    /**
-     * Resets the internal store
-     */
-    public function reset(): void
-    {
-        parent::reset();
-
-        $this->store["FROM"]      = "";
-        $this->store["RETURNING"] = [];
-    }
-
-    /**
-     * Adds the `RETURNING` clause
-     *
-     * @param array $columns
-     *
-     * @return Insert
-     */
-    public function returning(array $columns): self
-    {
-        $this->store["RETURNING"] = array_merge(
-            $this->store["RETURNING"],
-            $columns
+        return $this->connection->perform(
+            $this->getStatement(),
+            $this->getBindValues()
         );
-
-        return $this;
-    }
-
-    /**
-     * Sets a column = value condition
-     *
-     * @param string     $column
-     * @param mixed|null $value
-     *
-     * @return Insert
-     */
-    public function set(string $column, mixed $value = null): self
-    {
-        if (null === $value) {
-            $value = "NULL";
-        }
-
-        $this->store["COLUMNS"][$column] = $value;
-
-        $this->bind->remove($column);
-
-        return $this;
-    }
-
-    /**
-     * Builds the column list
-     *
-     * @return string
-     */
-    private function buildColumns(): string
-    {
-        $columns  = [];
-        $keyArray = array_keys($this->store["COLUMNS"]);
-        foreach ($keyArray as $column) {
-            $columns[] = $this->quoteIdentifier($column);
-        }
-
-        return " ("
-            . ltrim($this->indent($columns, ","))
-            . ") VALUES ("
-            . ltrim($this->indent(array_values($this->store["COLUMNS"]), ","))
-            . ")";
     }
 }
