--- tests/database/DataMapper/Query/QueryFactory/NewUpdateTest.php
@@ -1,36 +1,0 @@
-<?php
-
-/**
- * This file is part of the Phalcon Framework.
- *
- * (c) Phalcon Team <team@phalcon.io>
- *
- * For the full copyright and license information, please view the LICENSE.txt
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Phalcon\Tests\Database\DataMapper\Query\QueryFactory;
-
-use Phalcon\DataMapper\Query\QueryFactory;
-use Phalcon\DataMapper\Query\Update;
-use Phalcon\Tests\AbstractDatabaseTestCase;
-
-final class NewUpdateTest extends AbstractDatabaseTestCase
-{
-    /**
-     * Database Tests Phalcon\DataMapper\Query\QueryFactory :: newUpdate()
-     *
-     * @since  2020-01-20
-     *
-     * @group  common
-     */
-    public function testDmQueryQueryFactoryNewUpdate(): void
-    {
-        $connection = self::getDataMapperConnection();
-        $factory    = new QueryFactory();
-        $update     = $factory->newUpdate($connection);
-        $this->assertInstanceOf(Update::class, $update);
-    }
-}
